% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iGess}
\alias{iGess}
\title{IGESS}
\usage{
iGess(X, y, Z = NULL, SS = NULL, opts = NULL)
}
\arguments{
\item{X}{a N by P matrix of the genotype data, N-sample size, P - SNPs number}

\item{y}{a N by 1 matrix of the phenotypes}

\item{Z}{a N by M(number of covariates) matrix, the covariates of the N individuals, set to NULL by default}

\item{SS}{a K by P matrix for summary statistics, set to NULL by default}

\item{opts}{a list with fields(max_iter, display_gap), NULL corresponds to list(max_iter=600, display_gap = 60)}
}
\value{
List of model parameters, refer to \code{wrap_fit} function in Rcpp_auc.hpp for the fields
}
\description{
IGESS Model Training
}
\details{
\code{iGess} fits the iGess model. It requires to provide genotype matrix \code{X}
and phenotype matrix \code{y}, while users can also provide covariates data to \code{Z},
the summary statistics to \code{SS}, and the settings to \code{opts}.  these last three
parameters \code{Z}, \code{SS},\code{opts} could be set to NULL.
}
\examples{
##Working with no summary statistics, no covariates and options
data(IGESSDB)
fit <- iGess(Model$X,Model$y)


#Working with summary statistics(p-values) and no covariates
data(IGESSDB)
fit <- iGess(Model$X, Model$y, NULL, Model$AIPVal)

#Options Specified
data(IGESSDB)
opts = list(max_iter=600, display_gap = 100)
fit <- iGess(X,y,NULL,SS,opts)

}

