% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iGessCV}
\alias{iGessCV}
\title{IGESSCV}
\usage{
iGessCV(X, y, Z = NULL, SS = NULL, opts = NULL)
}
\arguments{
\item{X}{a N by P matrix of the genotype data, N-sample size, P - SNPs number}

\item{y}{a N by 1 matrix of the phenotypes}

\item{Z}{a N by M(number of covariates) matrix, the covariates of the N individuals, set to NULL by default}

\item{SS}{a K by P matrix for summary statistics, set to NULL by default}

\item{opts}{a list with fields(max_iter, display_gap), NULL corresponds to list(max_iter=600, display_gap = 60, n_fold = 5)}
}
\value{
list(auc, cor), auc is for the case-control studies, cor is for quantitative trait studies
}
\description{
evaluating the performance of prediction by cross validation
}
\details{
\code{iGessCV} evaluating the performance of prediction by cross validation.
It requires to provide genotype matrix \code{X}
and phenotype matrix \code{y}, while users can also provide covariates data to \code{Z},
the summary statistics to \code{SS}, and the settings to \code{opts}.  these last three
parameters \code{Z}, \code{SS},\code{opts} could be set to NULL.
}
\examples{
##Working with no summary statistics, no covariates and options
data(IGESSDB)
pair <- iGessCV(Model$X,Model$y)


#Working with summary statistics(p-values) and no covariates
data(IGESSDB)
pair <- iGessCV(Model$X, Model$y, NULL, Model$AIPVal)

#Options Specified
data(IGESSDB)
opts = list(max_iter=600, display_gap = 100)
pair <- iGessCV(X,y,NULL,SS,opts)

}

