% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iGessPredict}
\alias{iGessPredict}
\title{iGessPredict}
\usage{
iGessPredict(fit_, X, Z = NULL)
}
\arguments{
\item{X}{the given data matrix}

\item{Z}{covariates, set to be NULL by default}

\item{fit}{List of the IGESSfit fields}
}
\value{
y - predict of the phenotypes
}
\description{
predict the output of the given X and Covariates
}
\details{
\code{iGessPredict} predict the output of the given X and Covariates
}
\examples{


}

