% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(genoplinkfile, summaryfile_ = NULL, configfile_ = NULL)
}
\arguments{
\item{genoplinkfile}{the plink file for the genotype data}

\item{summaryfile}{the summary statistics file for the p-values}

\item{configfile}{the config file for summary statistics to specify the columns of pvalues}
}
\value{
list(X, y) - genotype data and its phenotypes or list(X,y,pvalues) - X, y and its corresponding pvalues
}
\description{
read the binary plink file or correpsonding summary statistics file into objects
}
\details{
\code{read_data} evaluating the performance of prediction by cross validation. It requires to provide binary plink file name to \code{genoplinkfile}, the binary file
should be SNP-major type in the current version of IGESS, while the users can also provide
summary statistics file name to \code{summaryfile},
the config file \code{configfile} is a must when \code{summaryfile} is provided.

\code{read_data} take the overlap of genotype data and summary statistics with respect to the SNPs
}
\examples{
# file name and its format
# the required binary files are CD.bed, CD.bim, CD.fam, \\code{genoplinkfile} should be \\code{CD} in its own directory.
# for \\code{summaryfile}, take the headers below for example
CHR SNP POS major_al minor_al wtccc_info narac1_info narac2_info eira_info canada_info brass_info
wtccc_z narac1_z narac2_z eira_z canada_z brass_z cases_MM cases_Mm cases_mm controls_MM controls_Mm
controls_mm meta_OR OR_95\%CI_lo OR_95\%CI_up meta_z meta_2tP CochranQ Q_Pval

#we should specify the column names for the summary statistics required and the column names for chromsome and the SNP name

#the content of \\code{configfile} is like following
zvalue=narac1_z,canada_z,brass_z
snp=SNP
chr=CHR

#zvalue should be replaced by pvalue if p-values are provided

#Working with no summary statistics
read_data(genoplinkfile)


#Working with summary statistics
read_data(genoplinkfile, summaryfile, configfile)


}

